Rename hyprland.async to hyprland.asyncio
