+++ README.md
@@ -7,28 +7,26 @@ An unofficial [type-safe](https://microsoft.github.io/pyright/), async API wrapp
 [examples/flicko.py](./examples/flicko.py)
 
 ```py
-from __future__ import annotations
-
 import asyncio
-
-from hyprland.binds import Key, Mod
-from hyprland.dispatch import (
+from hyprland import (
    Exec,
    Exit,
+   Key,
    KillActive,
+   Mod,
    MoveToWorkspace,
-   Workspace,
+   SwitchWorkspace,
    WorkspaceID,
 )
-from hyprland.socket import execute_batch_async
+from hyprland.asyncio import execute_batch
 
 
 async def main():
-   await execute_batch_async(
+   await execute_batch(
       (Mod.SUPER + Key("M")).bind(Exit()),
       (Mod.SUPER + Key("Return")).bind(Exec("wezterm")),
       (Mod.SUPER + Key("Q")).bind(KillActive()),
-      *((Mod.SUPER + Key(str(i))).bind(Workspace(WorkspaceID(i))) for i in range(1, 11)),
+      *((Mod.SUPER + Key(str(i))).bind(SwitchWorkspace(WorkspaceID(i))) for i in range(1, 11)),
       *((Mod.SUPER + Mod.SHIFT + Key(str(i))).bind(MoveToWorkspace(WorkspaceID(i))) for i in range(1, 11)),
    )
